Immersive: remove Axe.DoFuction patches
It appears that the Axe.DoFunction Harmony patch of the
ImmersiveSprinklers mod executes in may situations other than using the
actual Axe tools. In some cases, I have experienced it executing
continuously at basically every game time. The result is that sprinklers
and scarecrows immediately get picked up and returned to the player when
the mouse goes over them.

Its unclear why this patch causes this behavior, but it essentially
breaks the functionality of the mod. Based on logging it appears that
only the axe has this issue.

Since the pickaxe (and standard remove key) can still be used to pick up
sprinklers, remove the axe patches from both Immersive mods.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>