Immersive: add escape hatch for pickaxe pickup
The Harmony patch for the Axe.DoFunction caused gameplay issues in
situations where it ran even when the user did not provide the Axe tool
input. It is plausible the same issue may exist for the equivalent patch
to the pickaxe tool.

Make picking up with the pickaxe configurable via the options, so that
this can be disabled if it appears to happen in game. This can allow
disabling the option to save the game before further debugging of what
causes the function execute Unexpectedly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>